rename states.py to arrays.py
