+++ starsim/arrays.py
